+++ lib/rspec/matchers/dsl.rb
@@ -147,8 +147,14 @@ module RSpec
         # is rarely necessary, but can be helpful, for example, when specifying
         # asynchronous processes that require different timeouts.
         #
+        # By default the match block will swallow expectation errors (e.g.
+        # caused by using an expectation such as `expect(1).to eq 2`), if you
+        # with to allow these to bubble up, pass in the option
+        # `:notify_expectation_failures => true`.
+        #
+        # @param [Hash] options for defining the behavior of the match block.
         # @yield [Object] actual the actual value (i.e. the value wrapped by `expect`)
-        def match_when_negated(&match_block)
+        def match_when_negated(options={}, &match_block)
           define_user_override(:does_not_match?, match_block) do |actual|
             begin
               @actual = actual
@@ -156,6 +162,7 @@ module RSpec
                 super(*actual_arg_for(match_block))
               end
             rescue RSpec::Expectations::ExpectationNotMetError
+              raise if options[:notify_expectation_failures]
               false
             end
           end